+++ app/src/renderer/modules/pca/ComponentsCountPicker.tsx
