+++ app/src/renderer/modules/pca/Plot2D.tsx
