+++ app/src/renderer/modules/pca/PrincipalComponentsAnalysis.tsx
