+++ app/src/renderer/modules/pca/ScaleHandler.tsx
