+++ app/src/renderer/modules/pca/TargetAndFeaturesPicker.tsx
@@ -18,7 +18,7 @@ import {
 import CheckBoxIcon from '@mui/icons-material/CheckBox';
 import CheckBoxOutlineBlankIcon from '@mui/icons-material/CheckBoxOutlineBlank';
 
-import { AnalysisStepLogic } from '../analysis-step';
+import { AnalysisStepLogic } from '../../components/analysis-step';
 import { PCA } from './config';
 
 interface PossibleValues {
