+++ app/src/renderer/modules/pca/config/reducer.ts
