+++ app/src/renderer/modules/pca/index.tsx
